backend/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.5
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
executing==1.1.1
fastapi==0.85.1
fastjsonschema==2.16.2
firebase-admin==6.0.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
google-api-core==2.10.2
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
gspread==5.6.0
h11==0.14.0
httplib2==0.20.4
httptools==0.5.0
idna==3.4
image==1.5.33
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.24.2
oauthlib==3.2.2
orjson==3.8.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.8
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-slugify==6.1.2
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.3.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.3
stack-data==0.5.1
starlette==0.20.4
terminado==0.16.0
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.2
widgetsnbextension==4.0.3
qrcode==7.4.2
python-dotenv==1.0.0
